Move templates and tests
