+++ directory/tests/factories/__init__.py
@@ -1,0 +1,3 @@
+from directory.tests.factories.entry import *  # noqa: F403,F401
+from directory.tests.factories.pages import *  # noqa: F403,F401
+from directory.tests.factories.taxonomy import *  # noqa: F403,F401
