+++ directory/tests/factories/taxonomy.py
@@ -1,14 +1,6 @@
 import factory
-import wagtail_factories
 
-from directory.models import DirectoryPage, Language, Country, Topic
-
-
-class DirectoryPageFactory(wagtail_factories.PageFactory):
-    class Meta:
-        model = DirectoryPage
-    parent = factory.SubFactory(wagtail_factories.PageFactory, parent=None)
-    title = factory.Faker('sentence')
+from directory.models import Language, Country, Topic
 
 
 class LanguageFactory(factory.Factory):
